// src/api/generated/models/authorsListParams.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Bookora API
 * Contract OpenAPI chính thức cho backend Bookora. Frontend dùng cookie credentials, cookie accessToken/refreshToken và header X-CSRF-Token cho mutation cần CSRF.
 * OpenAPI spec version: 1.0
 */
import type { AuthorsListSortBy } from './authorsListSortBy';
import type { AuthorsListSortOrder } from './authorsListSortOrder';
import type { AuthorsListUsageStatus } from './authorsListUsageStatus';

export type AuthorsListParams = {
/**
 * @minimum 1
 */
page?: number;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
search?: string;
usageStatus?: AuthorsListUsageStatus;
createdFrom?: string;
createdTo?: string;
sortOrder?: AuthorsListSortOrder;
sortBy?: AuthorsListSortBy;
};